refactored package name to CLAS12Swim
