+++ common-tools/cnuphys/swimmer/src/main/java/cnuphys/CLAS12Swim/CLAS12RhoListener.java
@@ -1,16 +1,17 @@
-package cnuphys.dormandPrince;
+package cnuphys.CLAS12Swim;
 
 import java.io.File;
 import java.io.FileNotFoundException;
 
+import cnuphys.magfield.FastMath;
 import cnuphys.magfield.MagneticFieldInitializationException;
 import cnuphys.magfield.MagneticFields;
 import cnuphys.magfield.MagneticFields.FieldType;
 
-public class CLAS12ZListener extends CLAS12BoundaryListener {
+public class CLAS12RhoListener extends CLAS12BoundaryListener {
 	
-	//the target z (cm)
-	private double _zTarget;
+	//the target rho (cm)
+	private double _rhoTarget;
 	
 	//the starting sign. When this changes we have crossed.
 	private double _startSign;
@@ -18,21 +19,23 @@ public class CLAS12ZListener extends CLAS12BoundaryListener {
 	/**
 	 * Create a CLAS12 boundary target Z listener, for swimming to a fixed z
 	 * 
-	 * @param ivals           the initial values of the swim
-	 * @param zTarget         the target z (cm)
-	 * @param sMax            the final or max path length (cm)
-	 * @param accuracy        the desired accuracy (cm)
+	 * @param ivals         the initial values of the swim
+	 * @param rhoTarget     the target rho (cylindrical r) (cm)
+	 * @param sMax          the final or max path length (cm)
+	 * @param accuracy      the desired accuracy (cm)
 	 */
-	public CLAS12ZListener(CLAS12Values ivals, double zTarget, double sMax, double accuracy) {
+	public CLAS12RhoListener(CLAS12Values ivals, double rhoTarget, double sMax, double accuracy) {
 		super(ivals, sMax, accuracy);
-		_zTarget = zTarget;
-		_startSign = sign(ivals.z);
+		_rhoTarget = rhoTarget;
+		
+		double x = ivals.x;
+		double y = ivals.y;
+		_startSign = sign(Math.hypot(x, y));
 	}
 
 	@Override
 	public boolean crossedBoundary(double newS, double[] newU) {
-		double newZ = newU[2];
-		int sign = sign(newZ);
+		int sign = sign(rho(newU));
 		
 		if (sign != _startSign) {
 			return true;
@@ -40,15 +43,23 @@ public class CLAS12ZListener extends CLAS12BoundaryListener {
 		return false;
 	}
 	
+	//the rho (cylindrical r) of the state vector in cm
+	private double rho(double u[]) {
+		double x = u[0];
+		double y = u[1];
+		return FastMath.hypot(x, y);
+	}
+	
 	@Override
 	public boolean accuracyReached(double newS, double[] newU) {
-		double dZ = Math.abs(newU[2] - _zTarget);
-		return dZ < _accuracy;
+		double dRho = Math.abs(rho(newU) - _rhoTarget);
+//		System.err.println("dRho: " + dRho);
+		return dRho < _accuracy;
 	}
 
-	// left or right of the target Z?
-	private int sign(double z) {
-		return (z < _zTarget) ? -1 : 1;
+	// left or right of the target rho?
+	private int sign(double rho) {
+		return (rho < _rhoTarget) ? -1 : 1;
 	}
 	
 	/**
@@ -63,16 +74,16 @@ public class CLAS12ZListener extends CLAS12BoundaryListener {
 	public double interpolate(double s1, double[] u1, double s2, double[] u2, double u[]) {
 
 		//simple linear interpolation
-		double z1 = u1[2];
-		double z2 = u2[2];
+		double rho1 = rho(u1);
+		double rho2 = rho(u2);
 
 		// unlikely, but just in case
-		if (Math.abs(z2 - z1) < TINY) {
+		if (Math.abs(rho2 - rho1) < TINY) {
 			System.arraycopy(u2, 0, u, 0, 6);
 			return s2;
 		}
 
-		double t = (_zTarget - z1) / (z2 - z1);
+		double t = (_rhoTarget - rho1) / (rho2 - rho1);
 		double s = s1 + t * (s2 - s1);
 
 		for (int i = 0; i < 6; i++) {
@@ -82,6 +93,8 @@ public class CLAS12ZListener extends CLAS12BoundaryListener {
 		return s;
 
 	}
+
+	
 
 	// used for testing
 	public static void main(String arg[]) {
@@ -101,32 +114,32 @@ public class CLAS12ZListener extends CLAS12BoundaryListener {
 
 		System.out.println("Active Field Description: " + MagneticFields.getInstance().getActiveFieldDescription());
 
-		int q = -1;
+		
+		int q = 1;
 		double p = 2.0;
 		double theta = 15;
 		double phi = 5;
 		double xo = 0.01;
 		double yo = 0.02;
 		double zo = -0.01;
-		double ztarget = 575;
+		double rhotarget = 300;
 		double accuracy = 1.0e-5; //cm
 		double stepsizeAdaptive = accuracy/10; // starting stepsize in cm
 		double maxS = 800; // cm
 		double eps = 1.0e-6;
 
-		
 		MagneticFields.getInstance().setActiveField(FieldType.COMPOSITE);
-		CLAS12Swimmer clas12Swimmer = new CLAS12Swimmer(); //new
-		
+		CLAS12Swimmer clas12Swimmer = new CLAS12Swimmer(); // new
 
-		CLAS12SwimResult c12res = clas12Swimmer.swimZ(q, xo, yo, zo, p, theta, phi, ztarget, maxS, accuracy, stepsizeAdaptive, eps);
+		CLAS12SwimResult c12res = clas12Swimmer.swimRho(q, xo, yo, zo, p, theta, phi, rhotarget, maxS, accuracy,
+				stepsizeAdaptive, eps);
+
 		System.out.println("DP ACCURATE result:  " + c12res.toString() + "\n");
 
-// compare to interpolated approx
-		
-		c12res = clas12Swimmer.swimZInterp(q, xo, yo, zo, p, theta, phi, ztarget, maxS, stepsizeAdaptive, eps);
-		System.out.println("DP INTERP result:  " + c12res.toString() + "\n");
+		// compare to interpolated approx
 
+		c12res = clas12Swimmer.swimRhoInterp(q, xo, yo, zo, p, theta, phi, rhotarget, maxS, stepsizeAdaptive, eps);
+		System.out.println("DP INTERP result:  " + c12res.toString() + "\n");
 		
 	}
