+++ common-tools/cnuphys/swimmer/src/main/java/cnuphys/CLAS12Swim/CLAS12Swimmer.java
@@ -1,4 +1,4 @@
-package cnuphys.dormandPrince;
+package cnuphys.CLAS12Swim;
 
 import java.util.Hashtable;
