+++ common-tools/cnuphys/swimmer/src/main/java/cnuphys/CLAS12Swim/CLAS12ZListener.java
@@ -1,38 +1,38 @@
-package cnuphys.dormandPrince;
+package cnuphys.CLAS12Swim;
 
 import java.io.File;
 import java.io.FileNotFoundException;
 
-import cnuphys.adaptiveSwim.geometry.Plane;
 import cnuphys.magfield.MagneticFieldInitializationException;
 import cnuphys.magfield.MagneticFields;
 import cnuphys.magfield.MagneticFields.FieldType;
 
-public class CLAS12PlaneListener extends CLAS12BoundaryListener {
+public class CLAS12ZListener extends CLAS12BoundaryListener {
 	
-	//the target plane
-	private Plane _targetPlane;
+	//the target z (cm)
+	private double _zTarget;
 	
 	//the starting sign. When this changes we have crossed.
 	private double _startSign;
 
 	/**
-	 * Create a CLAS12 boundary target plane listener, for swimming to a fixed infinite plane
+	 * Create a CLAS12 boundary target Z listener, for swimming to a fixed z
 	 * 
 	 * @param ivals           the initial values of the swim
-	 * @param targetPlane     the target infinite plane
-	 * @param sMax         the final or max path length (cm)
+	 * @param zTarget         the target z (cm)
+	 * @param sMax            the final or max path length (cm)
 	 * @param accuracy        the desired accuracy (cm)
 	 */
-	public CLAS12PlaneListener(CLAS12Values ivals, Plane targetPlane, double sMax, double accuracy) {
+	public CLAS12ZListener(CLAS12Values ivals, double zTarget, double sMax, double accuracy) {
 		super(ivals, sMax, accuracy);
-		_targetPlane = targetPlane;
-		_startSign = _targetPlane.sign(ivals.x, ivals.y, ivals.z);
+		_zTarget = zTarget;
+		_startSign = sign(ivals.z);
 	}
 
 	@Override
 	public boolean crossedBoundary(double newS, double[] newU) {
-		int sign = _targetPlane.sign(newU[0], newU[1], newU[2]);
+		double newZ = newU[2];
+		int sign = sign(newZ);
 		
 		if (sign != _startSign) {
 			return true;
@@ -42,13 +42,17 @@ public class CLAS12PlaneListener extends CLAS12BoundaryListener {
 	
 	@Override
 	public boolean accuracyReached(double newS, double[] newU) {
-		double distance = _targetPlane.distance(newU[0], newU[1], newU[2]);
-		return distance < _accuracy;
+		double dZ = Math.abs(newU[2] - _zTarget);
+		return dZ < _accuracy;
 	}
 
+	// left or right of the target Z?
+	private int sign(double z) {
+		return (z < _zTarget) ? -1 : 1;
+	}
 	
 	/**
-	 * Interpolate between two points, one on each side of the plane
+	 * Interpolate between two points, one on each side of the boundary
 	 * @param s1 the path length of the "left" point (cm)
 	 * @param u1 the state vector of the "left" point
 	 * @param s2 the path length of the "right" point (cm)
@@ -57,14 +61,25 @@ public class CLAS12PlaneListener extends CLAS12BoundaryListener {
 	 * 
 	 */
 	public double interpolate(double s1, double[] u1, double s2, double[] u2, double u[]) {
-		
-		double t = _targetPlane.lineSegmentPlaneIntersection(u1, u2, u);
 
-		if (Double.isNaN(t)) {
-			return Double.NaN;
+		//simple linear interpolation
+		double z1 = u1[2];
+		double z2 = u2[2];
+
+		// unlikely, but just in case
+		if (Math.abs(z2 - z1) < TINY) {
+			System.arraycopy(u2, 0, u, 0, 6);
+			return s2;
 		}
-		
-		return s1 + t * (s2 - s1);
+
+		double t = (_zTarget - z1) / (z2 - z1);
+		double s = s1 + t * (s2 - s1);
+
+		for (int i = 0; i < 6; i++) {
+			u[i] = u1[i] + t * (u2[i] - u1[i]);
+		}
+
+		return s;
 
 	}
 
@@ -85,16 +100,15 @@ public class CLAS12PlaneListener extends CLAS12BoundaryListener {
 		}
 
 		System.out.println("Active Field Description: " + MagneticFields.getInstance().getActiveFieldDescription());
-		
-		Plane targetPlane = new Plane(1, 1, 1, 1, 1, 1);
 
 		int q = -1;
 		double p = 2.0;
-		double theta = 25;
+		double theta = 15;
 		double phi = 5;
 		double xo = 0.01;
 		double yo = 0.02;
 		double zo = -0.01;
+		double ztarget = 575;
 		double accuracy = 1.0e-5; //cm
 		double stepsizeAdaptive = accuracy/10; // starting stepsize in cm
 		double maxS = 800; // cm
@@ -105,17 +119,13 @@ public class CLAS12PlaneListener extends CLAS12BoundaryListener {
 		CLAS12Swimmer clas12Swimmer = new CLAS12Swimmer(); //new
 		
 
-		CLAS12SwimResult c12res = clas12Swimmer.swimPlane(q, xo, yo, zo, p, theta, phi, targetPlane, maxS, accuracy, stepsizeAdaptive, eps);
+		CLAS12SwimResult c12res = clas12Swimmer.swimZ(q, xo, yo, zo, p, theta, phi, ztarget, maxS, accuracy, stepsizeAdaptive, eps);
 		System.out.println("DP ACCURATE result:  " + c12res.toString() + "\n");
-		double[] u = c12res.getFinalU();
-		System.out.println("distance to plane: " + targetPlane.distance(u[0], u[1], u[2]) + " cm");
 
 // compare to interpolated approx
 		
-		c12res = clas12Swimmer.swimPlaneInterp(q, xo, yo, zo, p, theta, phi, targetPlane, maxS, stepsizeAdaptive, eps);
+		c12res = clas12Swimmer.swimZInterp(q, xo, yo, zo, p, theta, phi, ztarget, maxS, stepsizeAdaptive, eps);
 		System.out.println("DP INTERP result:  " + c12res.toString() + "\n");
-		u = c12res.getFinalU();
-		System.out.println("distance to plane: " + targetPlane.distance(u[0], u[1], u[2]) + " cm");
 
 		
 	}
