+++ common-tools/cnuphys/swimmer/src/main/java/cnuphys/CLAS12Swim/DormandPrince.java
@@ -1,4 +1,4 @@
-package cnuphys.dormandPrince;
+package cnuphys.CLAS12Swim;
 
 public class DormandPrince {
