+++ common-tools/cnuphys/swimmer/src/main/java/cnuphys/CLAS12Swim/ODE.java
@@ -1,4 +1,4 @@
-package cnuphys.dormandPrince;
+package cnuphys.CLAS12Swim;
 
 /**
  * Interface representing an ordinary differential equation (ODE).
