+++ common-tools/cnuphys/swimmer/src/main/java/cnuphys/CLAS12Swim/ODEStepListener.java
@@ -1,4 +1,4 @@
-package cnuphys.dormandPrince;
+package cnuphys.CLAS12Swim;
 
 /**
  * Interface for listening to steps taken by an ODE solver.
